- Extract TestRedisSpan to Redis package. - Change trace/redis abstraction to step on Client, not Conn.
