+++ trace/redis/redis.go
@@ -8,18 +8,32 @@ import (
 	"github.com/beatlabs/patron/trace"
 	"github.com/go-redis/redis/v7"
 	"github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/ext"
 )
 
-type connInfo struct {
-	instance string
-}
+const (
+	versionTag = "version"
+)
 
-func (c *connInfo) startSpan(ctx context.Context, opName, stmt string) (opentracing.Span, context.Context) {
-	return trace.RedisSpan(ctx, opName, trace.RedisComponent, trace.RedisDBType, stmt, c.instance)
-}
+var (
+	version = "dev"
+)
 
-// Client wraps redis.Client for easier usage.
-type Client redis.Client
+// RedisSpan starts a new Redis child span with specified tags.
+func RedisSpan(ctx context.Context, opName, cmp, dbType, instance, stmt string,
+	tags ...opentracing.Tag) (opentracing.Span, context.Context) {
+
+	sp, ctx := opentracing.StartSpanFromContext(ctx, opName)
+	ext.Component.Set(sp, cmp)
+	ext.DBType.Set(sp, dbType)
+	ext.DBInstance.Set(sp, instance)
+	ext.DBStatement.Set(sp, stmt)
+	for _, t := range tags {
+		sp.SetTag(t.Key, t.Value)
+	}
+	sp.SetTag(versionTag, version)
+	return sp, ctx
+}
 
 // Options wraps redis.Options for easier usage.
 type Options redis.Options
@@ -27,23 +41,23 @@ type Options redis.Options
 // Empty represents the error which is returned in case a key is not found.
 const Empty = redis.Nil
 
-// Conn represents a connection with a Redis client.
-type Conn struct {
-	connInfo
-	Client *redis.Client
+// Client represents a connection with a Redis client.
+type Client struct {
+	*redis.Client
 }
 
-// New returns a new connection to a Redis client.
-func New(ctx context.Context, opt Options) *Conn {
+func (c *Client) startSpan(ctx context.Context, opName, stmt string) (opentracing.Span, context.Context) {
+	return RedisSpan(ctx, opName, trace.RedisComponent, trace.RedisDBType, stmt, c.Options().Addr)
+}
+
+// New returns a new Redis client.
+func New(ctx context.Context, opt Options) *Client {
 	clientOptions := redis.Options(opt)
-	return &Conn{
-		connInfo{opt.Addr},
-		redis.NewClient(&clientOptions),
-	}
+	return &Client{redis.NewClient(&clientOptions)}
 }
 
 // Do creates and processes a custom Cmd on the underlying Redis client.
-func (c *Conn) Do(ctx context.Context, args ...interface{}) (interface{}, error) {
+func (c *Client) Do(ctx context.Context, args ...interface{}) (interface{}, error) {
 	sp, _ := c.startSpan(ctx, "redis.Do", fmt.Sprintf("%v", args))
 	cmd := c.Client.Do(args...)
 	trace.SpanComplete(sp, cmd.Err())
@@ -51,7 +65,7 @@ func (c *Conn) Do(ctx context.Context, args ...interface{}) (interface{}, error)
 }
 
 // Close closes the connection to the underlying Redis client.
-func (c *Conn) Close(ctx context.Context, args ...interface{}) error {
+func (c *Client) Close(ctx context.Context, args ...interface{}) error {
 	sp, _ := c.startSpan(ctx, "redis.Close", "")
 	cmd := c.Client.Close()
 	trace.SpanComplete(sp, cmd)
@@ -59,7 +73,7 @@ func (c *Conn) Close(ctx context.Context, args ...interface{}) error {
 }
 
 // Ping can be used to test whether a connection is still alive, or measure latency.
-func (c *Conn) Ping(ctx context.Context) (string, error) {
+func (c *Client) Ping(ctx context.Context) (string, error) {
 	sp, _ := c.startSpan(ctx, "redis.Ping", "")
 	cmd := c.Client.Ping()
 	trace.SpanComplete(sp, cmd.Err())
